requirements.txt
ultralytics==8.0.42
uvicorn[standard]==0.20.0
gunicorn==20.1.0
fastapi[all]==0.89.1
loguru
pytest
pytest-cov
python-dotenv
# Dependências para neuromarketing (opcionais - instale conforme necessário)
easyocr
pytesseract
transformers
torch
torchvision
deepface
opencv-python-headless
opencv-contrib-python-headless
scipy
numpy
pandas
pillow
# Dependências para análises avançadas de neuromarketing
mediapipe
scikit-image
scikit-learn